+++ packages/core/src/entity/EntityStore.ts
@@ -14,14 +14,32 @@ export class EntityStore extends EntityStoreBase {
   private _experienceEntry: Composition | undefined;
   private _unboundValues: CompositionUnboundValues | undefined;
 
-  constructor({ experienceEntry, entities, locale }: EntityStoreArgs) {
-    super({ entities, locale });
+  constructor(json: string);
+  constructor({ experienceEntry, entities, locale }: EntityStoreArgs);
 
-    if (isExperienceEntry(experienceEntry)) {
-      this._experienceEntry = (experienceEntry as ExperienceEntry).fields;
-      this._unboundValues = (experienceEntry as ExperienceEntry).fields.unboundValues;
+  constructor(options: string | EntityStoreArgs) {
+    if (typeof options === 'string') {
+      const data = JSON.parse(options);
+      const { _experienceEntry, _unboundValues, locale, entryMap, assetMap } = data.entityStore;
+      super({
+        entities: [
+          ...(Object.values(entryMap) as Entry[]),
+          ...(Object.values(assetMap) as Asset[]),
+        ],
+        locale,
+      });
+      this._experienceEntry = _experienceEntry;
+      this._unboundValues = _unboundValues;
     } else {
-      throw new Error('Provided entry is not experience entry');
+      const { experienceEntry, entities, locale } = options;
+      super({ entities, locale });
+
+      if (isExperienceEntry(experienceEntry)) {
+        this._experienceEntry = (experienceEntry as ExperienceEntry).fields;
+        this._unboundValues = (experienceEntry as ExperienceEntry).fields.unboundValues;
+      } else {
+        throw new Error('Provided entry is not experience entry');
+      }
     }
   }
 
@@ -53,16 +71,24 @@ export class EntityStore extends EntityStoreBase {
     return this._experienceEntry?.usedComponents ?? [];
   }
 
-  public updateUnboundValues(unboundValues: CompositionUnboundValues) {
-    this._unboundValues = { ...(this._unboundValues ?? {}), ...unboundValues };
+  /**
+   * Extend the existing set of unbound values with the ones from the assembly definition.
+   * When creating a new assembly out of a container, the unbound value keys are copied and
+   * thus the existing and the added ones have colliding keys. In the case of overlapping value
+   * keys, the ones from the experience overrule the ones from the assembly definition as
+   * the latter one is certainly just a default value while the other one is from the actual instance.
+   * @param unboundValues set of unbound values defined in the assembly definition
+   */
+  public addAssemblyUnboundValues(unboundValues: CompositionUnboundValues) {
+    this._unboundValues = { ...unboundValues, ...(this._unboundValues ?? {}) };
   }
 
   public getValue(
     entityLinkOrEntity: UnresolvedLink<'Entry' | 'Asset'> | Entry | Asset,
-    path: string[]
+    path: string[],
   ): string | undefined {
     const isLink = (
-      entity: typeof entityLinkOrEntity
+      entity: typeof entityLinkOrEntity,
     ): entity is UnresolvedLink<'Entry' | 'Asset'> => entityLinkOrEntity.sys.type === 'Link';
 
     let entity: Entry | Asset;
@@ -74,7 +100,7 @@ export class EntityStore extends EntityStoreBase {
 
       if (!resolvedEntity || resolvedEntity.sys.type !== entityLinkOrEntity.sys.linkType) {
         console.warn(
-          `Experience references unresolved entity: ${JSON.stringify(entityLinkOrEntity)}`
+          `Experience references unresolved entity: ${JSON.stringify(entityLinkOrEntity)}`,
         );
         return;
       }